Rename modbus-utils package to modbus-client package
