+++ bin/modbus_client_rpc
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
 
 import sys
-from modbus_utils_rpc.main import main
+from modbus_client_rpc.main import main
 sys.exit(main(sys.argv))
